Fix api-report.test
