+++ apps/studio/__mocks__/hooks/analytics/useLogsQuery.ts
@@ -1,0 +1,2 @@
+const useLogsQuery = jest.fn().mockReturnValue([])
+export default useLogsQuery
