+++ apps/studio/__mocks__/lib/common/fetch.js
@@ -1,3 +1,3 @@
-export { isResponseOk } from 'lib/common/fetch'
 export const get = jest.fn()
 export const post = jest.fn()
+export const isResponseOk = jest.fn().mockReturnValue(true)
