+++ apps/studio/jest.config.mjs
@@ -16,6 +16,10 @@ const config = {
     'react-markdown': '<rootDir>/__mocks__/react-markdown.js',
     'sse.js': '<rootDir>/__mocks__/sse.js',
     'react-dnd': '<rootDir>/__mocks__/react-dnd.js',
+    // [Joshen] There's bound to be a better way to do this and we'll need to figure this out
+    'lib/common/fetch': '<rootDir>/__mocks__/lib/common/fetch',
+    'hooks/analytics/useLogsQuery': '<rootDir>/__mocks__/hooks/analytics/useLogsQuery',
+    'data/reports/api-report-query': '<rootDir>/__mocks__/hooks/useApiReport',
   },
   testEnvironment: 'jsdom',
   testTimeout: 10000,
