+++ apps/studio/tests/pages/projects/reports/api-report.test.js
@@ -1,21 +1,9 @@
-import { get } from 'lib/common/fetch'
-import { render } from '../../../helpers'
-import { waitFor, screen } from '@testing-library/react'
+import { screen } from '@testing-library/react'
+
 import { ApiReport } from 'pages/project/[ref]/reports/api-overview'
-import userEvent from '@testing-library/user-event'
+import { render } from '../../../helpers'
 
-// [Joshen] Am temporarily commenting out the breaking tests due to:
-// "TypeError: _fetch.get.mockReset is not a function" error from Jest
-// just so we get our jest unit/UI tests up and running first
-// Need to figure out how to mock the "get" method from lib/common/fetch properly
-
-beforeEach(() => {
-  // reset mocks between tests
-  // get.mockReset()
-  // get.mockImplementation(async (_url) => [{ result: [] }])
-})
-
-test(`static elements`, async () => {
+test(`Render static elements`, async () => {
   render(<ApiReport />)
   await screen.findByText('Total Requests')
   await screen.findByText('Response Errors')
@@ -26,51 +14,16 @@ test(`static elements`, async () => {
   await screen.findByText(/All Requests/)
 })
 
-test('refresh button', async () => {
+test('Render Total Requests section', async () => {
   render(<ApiReport />)
-  // await waitFor(() => expect(get).toBeCalled())
-  // get.mockReset()
-  // userEvent.click(await screen.findByText(/Refresh/))
-  // await waitFor(() => expect(get).toBeCalled())
+  await screen.findAllByText('/rest/v1/')
+  await screen.findAllByText('GET')
+  await screen.findAllByText('200')
 })
 
-test('append - api request routes', async () => {
-  //   get.mockImplementation(async (url) => {
-  //     if (decodeURIComponent(url).includes('request.path')) {
-  //       return {
-  //         result: [
-  //           { path: 'mypath', method: 'GET', status_code: 200, search: 'some-query', count: 22 },
-  //         ],
-  //       }
-  //     }
-  //     return { result: [{ timestamp: new Date().toISOString(), count: 123 }] }
-  //   })
-  //   render(<ApiReport />)
-  //   await waitFor(() => expect(get).toBeCalled())
-  //   await screen.findAllByText(/mypath/)
-  //   await screen.findAllByText(/GET/)
-  //   await screen.findAllByText(/200/)
-  //   await screen.findAllByText(/some\-query/)
-  //   await screen.findAllByText(/22/)
-})
-
-test('append - error routes', async () => {
-  //   get.mockImplementation(async (url) => {
-  //     const uri = decodeURIComponent(url)
-  //     if (uri.includes('400')) {
-  //       return {
-  //         result: [
-  //           { path: 'mypath', method: 'GET', status_code: 200, search: 'some-query', count: 22 },
-  //         ],
-  //       }
-  //     }
-  //     return { result: [{ timestamp: new Date().toISOString(), count: 123 }] }
-  //   })
-  //   render(<ApiReport />)
-  //   await waitFor(() => expect(get).toBeCalled())
-  //   await screen.findAllByText(/mypath/)
-  //   await screen.findAllByText(/GET/)
-  //   await screen.findAllByText(/200/)
-  //   await screen.findAllByText(/some\-query/)
-  //   await screen.findAllByText(/22/)
+test('Render Response Errors section', async () => {
+  render(<ApiReport />)
+  await screen.findAllByText('/auth/v1/user')
+  await screen.findAllByText('GET')
+  await screen.findAllByText('403')
 })
